+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-02T14:35:08.020Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-04T12:58:16.632Z")
     r"""The date and time the resource was last updated"""